Fixed model relation, added nested attributes to recipes

+++ db/migrate/20170916151313_create_ingredients.rb
@@ -2,7 +2,7 @@ class CreateIngredients < ActiveRecord::Migration[5.0]
   def change
     create_table :ingredients do |t|
       t.string :name
-      t.integer :package_quantity
+      t.float :package_quantity
       t.float :cost
 
       t.timestamps
